Continuous benchmarking: automation and organization
* Re-organize benchmarks; split into continuous and excluded
* Some small additions and fixes for benchmark code; to be extended
* Add tox config; disable numba boundschecks etc.
