+++ .github/workflows/cb.yml
@@ -1,0 +1,74 @@
+name: benchmarks
+
+on:
+    # trigger benchmark runs by adding a label, which is limited to users
+    # that can write to the repository:
+    pull_request_target:
+        types: [labeled]
+    # additionally trigger on pushes, meaning when PRs are merged:
+    push:
+
+jobs:
+    benchmarks:
+        runs-on: ubuntu-latest
+        if: >-
+            (github.repository_owner == 'LiberTEM') &&
+            (
+             (github.event_name == 'push') ||
+             (github.event_name == 'schedule') ||
+             (
+              (github.event_name == 'pull_request_target') &&
+              (contains(github.event.pull_request.labels.*.name, 'benchmark'))
+             )
+            )
+        steps:
+            - name: Remove benchmark label (so it can be re-triggered afterwards)
+              if: ${{ github.event_name == 'pull_request_target' }}
+              run: |
+               curl --silent --fail-with-body \
+                  -X DELETE \
+                  -H 'Accept: application/vnd.github.v3+json' \
+                  -H 'Authorization: token ${{ github.token }}' \
+                  'https://api.github.com/repos/${{ github.repository }}/issues/${{ github.event.number }}/labels/benchmark'
+            - name: Checkout
+              uses: actions/checkout@v4
+            - name: Mirror and get status
+              uses: jakob-fritz/github2lab_action@main
+              env:
+                MODE: 'all'
+                GITLAB_TOKEN: ${{ secrets.GITLAB_TOKEN }}
+                FORCE_PUSH: "true"  # ensure github and gitlab are in sync
+                GITLAB_HOSTNAME: "gitlab.jsc.fz-juelich.de"
+                GITLAB_PROJECT_ID: "7702"
+                GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+            - name: Setup Python
+              uses: actions/setup-python@v5
+              with:
+                python-version: '3.12'
+            - name: Compare benchmark results
+              if: github.event.pull_request
+              env:
+                GH_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+                PR: ${{ github.event.pull_request.html_url }}            
+              run: >-
+                wget https://raw.githubusercontent.com/LiberTEM/benchmarks/refs/heads/main/utils/compare.py &&
+                python3 -m venv venv &&
+                ./venv/bin/pip install click scipy &&
+                ./venv/bin/python3 compare.py ./artifacts/juwels_cpu_reference_run.zip ./artifacts/juwels_cpu.zip | gh pr comment $PR -F -
+            - name: Trigger artifact collection
+              if: ${{ github.event_name == 'push' }}
+              env:
+                TRIGGER_TOKEN: ${{ secrets.TRIGGER_BENCH_GH_TOKEN }}
+                RUN_ID: ${{ github.run_id }}
+              run: |
+                curl -X POST \
+                  -H "Accept: application/vnd.github+json" \
+                  -H "Authorization: Bearer $TRIGGER_TOKEN" \
+                  https://api.github.com/repos/LiberTEM/benchmarks/actions/workflows/collect-and-render.yml/dispatches \
+                  -d "{
+                    \"ref\": \"main\",
+                    \"inputs\": {
+                      \"run_id\": \"$RUN_ID\",
+                      \"source_repo\": \"LiberTEM/LiberTEM\"
+                    }
+                  }"
