--- benchmarks/io/data/.gitignore
@@ -1,1 +1,0 @@
-./*
