+++ tox.ini
@@ -123,11 +123,23 @@ passenv=
 
 [testenv:benchmark]
 changedir={toxinidir}
+setenv=
+    NUMBA_FULL_TRACEBACKS=0
+    NUMBA_BOUNDSCHECK=0
 commands=
     pytest --benchmark-enable --benchmark-warmup=on --benchmark-autosave --benchmark-storage "{toxinidir}/benchmark_results" -m "not compilation" {posargs:benchmarks/}
     pytest --benchmark-enable --benchmark-warmup=off --benchmark-autosave --benchmark-storage "{toxinidir}/benchmark_results" -m "compilation" {posargs:benchmarks/}
 
-[testenv:benchmark-cuda{101,102,110,114}]
+
+[testenv:benchmark-cb]
+changedir={toxinidir}
+setenv=
+    NUMBA_FULL_TRACEBACKS=0
+    NUMBA_BOUNDSCHECK=0
+commands=
+    pytest --benchmark-enable --benchmark-warmup=on --benchmark-save-data --benchmark-json=bench-results.json {posargs:benchmarks/continuous/}
+
+[testenv:benchmark-cuda{101,102,110,114,11x,12x}]
 changedir={toxinidir}
 commands=
     {[testenv:benchmark]commands}
